+++ test/unit/association_test.rb
@@ -769,6 +769,91 @@ class AssociationTest < MiniTest::Test
     Specified.where(foo_id: 1).create
   end
 
+  def test_get_with_relationship_for_model_with_custom_type
+    stub_request(:get, "http://example.com/document_users/1?include=file")
+        .to_return(headers: {content_type: "application/vnd.api+json"}, body: {
+            data: [
+                {
+                    id: '1',
+                    type: 'document_users',
+                    attributes: {
+                        name: 'John Doe'
+                    },
+                    relationships: {
+                        file: {
+                            links: {
+                                self: 'http://example.com/document_users/1/relationships/file',
+                                related: 'http://example.com/document_users/1/file'
+                            },
+                            data: {
+                                id: '2',
+                                type: 'document--files'
+                            }
+                        }
+                    }
+                }
+            ],
+            included: [
+                {
+                    id: '2',
+                    type: 'document--files',
+                    attributes: {
+                        url: 'http://example.com/downloads/2.pdf'
+                    }
+                }
+            ]
+        }.to_json)
+
+    user = DocumentUser.includes('file').find(1).first
+
+    assert_equal 'document--files', user.file.type
+    assert user.file.is_a?(DocumentFile)
+  end
+
+  def test_get_with_defined_relationship_for_model_with_custom_type
+    stub_request(:get, "http://example.com/document_stores/1?include=files")
+        .to_return(headers: {content_type: "application/vnd.api+json"}, body: {
+            data: [
+                {
+                    id: '1',
+                    type: 'document_stores',
+                    attributes: {
+                        name: 'store #1'
+                    },
+                    relationships: {
+                        files: {
+                            links: {
+                                self: 'http://example.com/document_stores/1/relationships/files',
+                                related: 'http://example.com/document_stores/1/files'
+                            },
+                            data: [
+                                {
+                                    id: '2',
+                                    type: 'document--files'
+                                }
+                            ]
+                        }
+                    }
+                }
+            ],
+            included: [
+                {
+                    id: '2',
+                    type: 'document--files',
+                    attributes: {
+                        url: 'http://example.com/downloads/2.pdf'
+                    }
+                }
+            ]
+        }.to_json)
+
+    user = DocumentStore.includes('files').find(1).first
+
+    assert_equal 1, user.files.size
+    assert_equal 'document--files', user.files.first.type
+    assert user.files.first.is_a?(DocumentFile)
+  end
+
   def test_load_include_from_dataset
     stub_request(:get, 'http://example.com/employees?include=chief&page[per_page]=2')
         .to_return(